+++ src/components/dialog.js
@@ -16,10 +16,8 @@
       setIsOpen(isVisible);
     }, [isVisible]);
 
-    useEffect(() => {
-      B.defineFunction('OpenDialog', openDialog);
-      B.defineFunction('CloseDialog', closeDialog);
-    }, []);
+    B.defineFunction('OpenDialog', openDialog);
+    B.defineFunction('CloseDialog', closeDialog);
 
     const isEmpty = !children.length;